Correção.

+++ TesteTopDown/Controllers/TaskController.cs
@@ -11,12 +11,13 @@ namespace TesteTopDown.Controllers
     [Authorize]
     [Route("[controller]")]
 
-    public class TaskController : ControllerBase
+    public class TaskController : MainController
     {
         private readonly ITaskService _taskService;
         private readonly ITaskRepository _taskRepository;
         private readonly IMapper _mapper;
-        public TaskController(ITaskService taskService, IMapper mapper, ITaskRepository taskRepository)
+        public TaskController(ITaskService taskService, IMapper mapper, 
+                                ITaskRepository taskRepository, INotificador notificador): base(notificador)
         {
             _taskService = taskService;
             _taskRepository = taskRepository;
@@ -25,34 +26,51 @@ namespace TesteTopDown.Controllers
         
         [HttpPost("Adicionar")]
         public async Task<IActionResult> AdicionarTarefa(AddTaskInput input)
-        {          
+        {
+            if (!ModelState.IsValid) return CustomResponse(ModelState);
 
             await _taskService.Adicionar(_mapper.Map<Tarefa>(input));
 
-            return Created("Tarefa criada com sucesso.", input);
+            return CustomResponse(input);
         }
-
+        
         [HttpPut("Atualizar")]
-        public IActionResult UpdateTask(AddTaskInput input)
+        public async Task<IActionResult> UpdateTask(AddTaskInput input)
         {
-            _taskService.Atualizar(_mapper.Map<Tarefa>(input));
-            return Ok("Tarefa Atualizada com sucesso.");
+            if (!ModelState.IsValid) return CustomResponse(ModelState);
+
+
+            await _taskService.Atualizar(_mapper.Map<Tarefa>(input));
+
+            return CustomResponse(input);
         }
 
         [HttpDelete("Remover")]
-        public IActionResult DeleteTask(int id)
+        public async Task<ActionResult<AddTaskInput>> DeleteTask(int id)
         {
-            _taskService.Remover(id);
-            return Ok("Tarefa Excluida com sucesso.");
+            var task = _mapper.Map<AddTaskInput>(await _taskRepository.ObterPorId(id));
+
+            if (task == null) return NotFound();
+
+            await  _taskService.Remover(id);
+
+            return  CustomResponse(task);
         }
 
         [AllowAnonymous]
-        [HttpGet("Consultar/{id:int}")]
-        public async Task<Tarefa> ObterTask(int id)
+        [HttpGet("Consultar/Id")]
+        public async Task<AddTaskInput> ObterTask(int id)
         {
-            return _mapper.Map<Tarefa>(_taskRepository.ObterPorId(id));            
+            return _mapper.Map<AddTaskInput>(await _taskRepository.ObterPorId(id));            
         }
 
+        [AllowAnonymous]
+        [HttpGet]
+        public async Task<IEnumerable<AddTaskInput>> ObterTodos()
+        {
+            var fornecedor = _mapper.Map<IEnumerable<AddTaskInput>>(await _taskRepository.ObterTodos());
+            return fornecedor;
+        }
 
     }
